03_etl/requirements.txt
python-dotenv==0.19.2
psycopg2-binary==2.9.3
elasticsearch==7.7.0